+++ sockeye/layers.py
@@ -13,7 +13,7 @@
 
 import logging
 import math
-from typing import Optional, Tuple, Union
+from typing import Dict, Optional, Tuple, Union
 
 import mxnet as mx
 import numpy as np
@@ -212,55 +212,62 @@ class OutputLayer:
                                     flatten=False)
 
 
-def split_heads(x: mx.sym.Symbol, length: int, heads: int) -> mx.sym.Symbol:
+def split_heads(x: mx.sym.Symbol, depth_per_head: int, heads: int) -> mx.sym.Symbol:
     """
     Returns a symbol with head dimension folded into batch and depth divided by the number of heads.
 
     :param x: Symbol of shape (batch, length, depth).
-    :param length: Sequence length.
+    :param depth_per_head: Depth per head.
     :param heads: Number of heads.
     :return: Symbol of shape (batch * heads, length, depth_per_heads).
     """
     # (batch, length, heads, depth_per_head)
-    x = mx.sym.reshape(data=x, shape=(0, length, heads, -1))
+    x = mx.sym.reshape(data=x, shape=(0, -1, heads, depth_per_head))
     # (batch, heads, length, depth/heads)
     x = mx.sym.transpose(data=x, axes=(0, 2, 1, 3))
     # (batch * heads, length, depth/heads)
-    return mx.sym.reshape(data=x, shape=(-3, length, -1))
+    return mx.sym.reshape(data=x, shape=(-3, -1, depth_per_head))
 
 
-def combine_heads(x: mx.sym.Symbol, length: int, heads: int) -> mx.sym.Symbol:
+def combine_heads(x: mx.sym.Symbol, depth_per_head: int, heads: int) -> mx.sym.Symbol:
     """
     Returns a symbol with both batch & length, and head & depth dimensions combined.
 
     :param x: Symbol of shape (batch * heads, length, depth_per_head).
-    :param length: Sequence length.
+    :param depth_per_head: Depth per head.
     :param heads: Number of heads.
     :return: Symbol of shape (batch, length, depth).
     """
     # (batch, heads, length, depth_per_head)
-    x = mx.sym.reshape(data=x, shape=(-4, -1, heads, length, 0))
+    x = mx.sym.reshape(data=x, shape=(-4, -1, heads, 0, depth_per_head))
     # (batch, length, heads, depth_per_head)
     x = mx.sym.transpose(x, axes=(0, 2, 1, 3))
     # (batch, length, depth)
-    return mx.sym.reshape(x, shape=(-1, length, -3))
+    return mx.sym.reshape(x, shape=(-1, 0, depth_per_head * heads))
 
 
-def broadcast_to_heads(x: mx.sym.Symbol, heads: int) -> mx.sym.Symbol:
+def broadcast_to_heads(x: mx.sym.Symbol, num_heads: int, ndim: int, fold_heads: bool = True) -> mx.sym.Symbol:
     """
-    Broadcasts a 1d vector of shape (batch,) to (batch*heads, 1).
+    Broadcasts batch-major input of shape (batch, d1 ... dn-1) to (batch*heads, d1 ... dn-1).
 
-    :param x: Symbol(batch,)
-    :param heads: Number of heads.
-    :return: Symbol(batch * heads, 1)
+    :param x: Batch-major input. Shape: (batch, d1 ... dn-1).
+    :param num_heads: Number of heads.
+    :param ndim: Number of dimensions in x.
+    :param fold_heads: Whether to fold heads dimension into batch dimension.
+    :return: Tensor with each sample repeated heads-many times.
+             Shape: (batch * heads, d1 ... dn-1) if fold_heads == True, (batch, heads, d1 ... dn-1) else.
     """
+    dims = [0] * (ndim - 1)
     # x: (batch, 1)
     x = mx.sym.expand_dims(x, axis=1)
-    # x: (batch, heads)
-    x = mx.sym.broadcast_to(x, shape=(0, heads))
-    # x: (batch * heads, 1)
-    x = mx.sym.reshape(x, shape=(-3,))
-    return x
+    # x: (batch, heads, dims...)
+    x = mx.sym.broadcast_to(x, shape=[0, num_heads] + dims)
+    if fold_heads:
+        # (batch * heads, dims...)
+        return mx.sym.reshape(x, shape=[-3] + dims)
+    else:
+        # x: (batch, heads, dims...)
+        return x
 
 
 def dot_attention(queries: mx.sym.Symbol,
@@ -268,7 +275,8 @@ def dot_attention(queries: mx.sym.Symbol,
                   values: mx.sym.Symbol,
                   lengths: Optional[mx.sym.Symbol] = None,
                   dropout: float = 0.0,
-                  bias: Optional[mx.sym.Symbol] = None):
+                  bias: Optional[mx.sym.Symbol] = None,
+                  prefix: Optional[str] = ''):
     """
     Computes dot attention for a set of queries, keys, and values.
 
@@ -277,14 +285,15 @@ def dot_attention(queries: mx.sym.Symbol,
     :param values: Attention values. Shape: (n, lk, dv).
     :param lengths: Optional sequence lengths of the keys. Shape: (n,).
     :param dropout: Dropout probability.
-    :param bias: Optional bias tensor. Shape: (1, lq, lk).
+    :param bias: Optional 3d bias tensor.
+    :param prefix: Optional prefix
     :return: 'Context' vectors for each query. Shape: (n, lq, dv).
     """
     utils.check_condition(lengths is not None or bias is not None,
                           "Must provide either length or bias argument for masking")
 
     # (n, lq, lk)
-    logits = mx.sym.batch_dot(lhs=queries, rhs=keys, transpose_b=True)
+    logits = mx.sym.batch_dot(lhs=queries, rhs=keys, transpose_b=True, name='%sdot' % prefix)
 
     if lengths is not None:
         # mask lk dimension
@@ -293,18 +302,18 @@ def dot_attention(queries: mx.sym.Symbol,
         logits = mx.sym.SequenceMask(data=logits,
                                      use_sequence_length=True,
                                      sequence_length=lengths,
-                                     value=-99999999.)
+                                     value=C.LARGE_NEGATIVE_VALUE)
         # (n, lq, lk)
         logits = mx.sym.transpose(data=logits, axes=(1, 2, 0))
 
     if bias is not None:
-        logits = mx.sym.broadcast_add(logits, bias)
+        logits = mx.sym.broadcast_add(logits, bias, name='%sbias_add' % prefix)
 
     probs = mx.sym.softmax(logits, axis=-1)
     probs = mx.sym.Dropout(probs, p=dropout) if dropout > 0.0 else probs
 
     # (n, lq, lk) x (n, lk, dv) -> (n, lq, dv)
-    return mx.sym.batch_dot(lhs=probs, rhs=values)
+    return mx.sym.batch_dot(lhs=probs, rhs=values, name='%scontexts' % prefix)
 
 
 class MultiHeadAttentionBase:
@@ -339,26 +348,35 @@ class MultiHeadAttentionBase:
                 queries: mx.sym.Symbol,
                 keys: mx.sym.Symbol,
                 values: mx.sym.Symbol,
-                queries_max_length: int,
-                memory_max_length: int,
                 lengths: Optional[mx.sym.Symbol] = None,
                 bias: Optional[mx.sym.Symbol] = None) -> mx.sym.Symbol:
+        """
+        Returns context vectors of multi-head dot attention.
+
+        :param queries: Query tensor. Shape: (batch_size, query_max_length, depth).
+        :param keys: Keys. Shape: (batch_size, memory_max_length, depth).
+        :param values: Values. Shape: (batch_size, memory_max_length, depth).
+        :param lengths: Optional lengths of keys. Shape: (batch_size,).
+        :param bias: Optional 3d bias.
+        :return: Context vectors. Shape: (batch_size, query_max_length, output_depth).
+        """
         # scale by sqrt(depth_per_head)
         queries = queries * (self.depth_per_head ** -0.5)
 
         # (batch*heads, length, depth/heads)
-        queries = split_heads(queries, queries_max_length, self.heads)
-        keys = split_heads(keys, memory_max_length, self.heads)
-        values = split_heads(values, memory_max_length, self.heads)
-        lengths = broadcast_to_heads(lengths, self.heads) if lengths is not None else lengths
+        queries = split_heads(queries, self.depth_per_head, self.heads)
+        keys = split_heads(keys, self.depth_per_head, self.heads)
+        values = split_heads(values, self.depth_per_head, self.heads)
+        lengths = broadcast_to_heads(lengths, self.heads, ndim=1, fold_heads=True) if lengths is not None else lengths
 
-        # (batch*heads, queries_max_length, depth_per_head)
-        contexts = dot_attention(queries, keys, values, lengths=lengths, dropout=self.dropout, bias=bias)
+        # (batch*heads, query_max_length, depth_per_head)
+        contexts = dot_attention(queries, keys, values,
+                                 lengths=lengths, dropout=self.dropout, bias=bias, prefix=self.prefix)
 
-        # (batch, queries_max_length, depth)
-        contexts = combine_heads(contexts, queries_max_length, self.heads)
+        # (batch, query_max_length, depth)
+        contexts = combine_heads(contexts, self.depth_per_head, self.heads)
 
-        # contexts: (batch, queries_max_length, output_depth)
+        # contexts: (batch, query_max_length, output_depth)
         contexts = mx.sym.FullyConnected(data=contexts,
                                          weight=self.w_h2o,
                                          bias=self.b_h2o,
@@ -386,22 +404,25 @@ class MultiHeadSelfAttention(MultiHeadAttentionBase):
                  depth_out: int = 512,
                  dropout: float = 0.0) -> None:
         super().__init__(prefix, depth_att, heads, depth_out, dropout)
-
         self.w_i2h = mx.sym.Variable("%si2h_weight" % prefix)
         self.b_i2h = mx.sym.Variable("%si2h_bias" % prefix)
 
     def __call__(self,
                  inputs: mx.sym.Symbol,
-                 max_length: int,
-                 lengths: Optional[mx.sym.Symbol] = None,
-                 bias: Optional[mx.sym.Symbol] = None) -> mx.sym.Symbol:
+                 input_lengths: Optional[mx.sym.Symbol] = None,
+                 bias: Optional[mx.sym.Symbol] = None,
+                 cache: Optional[Dict[str, Optional[mx.sym.Symbol]]] = None) -> mx.sym.Symbol:
         """
+        Computes multi-head attention on a set of inputs, serving as queries, keys, and values.
+        If sequence lengths are provided, they will be used to mask the attention scores.
+        A bias mask may also be used to mask the attention scores.
+        May also use a cache of previously computed inputs.
         Returns a symbol of shape (batch, max_length, output_depth).
 
-        :param inputs: Symbol of shape (batch, max_length, input_depth).
-        :param max_length: Size of time dimension.
-        :param lengths: Optional lengths of inputs. Symbol of shape (batch, 1).
-        :param bias: Optional (auto-regressive bias). Symbol of shape (1, max_length, max_length).
+        :param inputs: Input Data. Shape: (batch, max_length, input_depth).
+        :param input_lengths: Optional lengths of inputs to mask attention scores. Shape: (batch, 1).
+        :param bias: Optional 3d bias tensor to mask attention scores.
+        :param cache: Optional dictionary of previously computed keys and values.
         :return: Symbol of shape (batch, max_length, output_depth).
         """
         # combined: (batch, max_length, depth * 3)
@@ -416,12 +437,15 @@ class MultiHeadSelfAttention(MultiHeadAttentionBase):
         # pylint: disable=unbalanced-tuple-unpacking
         queries, keys, values = mx.sym.split(data=combined, num_outputs=3, axis=2)
 
+        if cache is not None:
+            # append new keys & values to cache, update the cache
+            keys = cache['k'] = keys if cache['k'] is None else mx.sym.concat(cache['k'], keys, dim=1)
+            values = cache['v'] = values if cache['v'] is None else mx.sym.concat(cache['v'], values, dim=1)
+
         return self._attend(queries,
                             keys,
                             values,
-                            queries_max_length=max_length,
-                            memory_max_length=max_length,
-                            lengths=lengths,
+                            lengths=input_lengths,
                             bias=bias)
 
 
@@ -450,19 +474,20 @@ class MultiHeadAttention(MultiHeadAttentionBase):
 
     def __call__(self,
                  queries: mx.sym.Symbol,
-                 queries_max_length: int,
                  memory: mx.sym.Symbol,
-                 memory_lengths: mx.sym.Symbol,
-                 memory_max_length: int) -> mx.sym.Symbol:
+                 memory_lengths: Optional[mx.sym.Symbol] = None,
+                 bias: Optional[mx.sym.Symbol] = None) -> mx.sym.Symbol:
         """
+        Computes multi-head attention for queries given a memory tensor.
+        If sequence lengths are provided, they will be used to mask the attention scores.
+        A bias mask may also be used to mask the attention scores.
         Returns a symbol of shape (batch, max_length, output_depth).
 
-        :param queries: Symbol of shape (batch, queries_max_length, input_depth). TODO?
-        :param queries_max_length: Size of queries time dimension.
-        :param memory: Symbol of shape (batch, memory_max_length, input_depth).
-        :param memory_lengths: Symbol of shape (batch, 1).
-        :param memory_max_length: Size of memory time dimension.
-        :return: Symbol of shape (batch, queries_max_length, output_depth).
+        :param queries: Query tensor. Shape: (batch, query_max_length, input_depth).
+        :param memory: Memory data to attend to. Shape: (batch, memory_max_length, input_depth).
+        :param memory_lengths: Optional lengths of memory to mask attention scores. Shape: (batch, 1).
+        :param bias: Optional 3d bias tensor to mask attention scores.
+        :return: Symbol of shape (batch, query_seq_len, output_depth).
         """
         # (batch, memory_max_length, depth * 2)
         combined = mx.sym.FullyConnected(data=memory,
@@ -478,21 +503,17 @@ class MultiHeadAttention(MultiHeadAttentionBase):
         # pylint: disable=unbalanced-tuple-unpacking
         keys, values = mx.sym.split(data=combined, num_outputs=2, axis=2)
 
-        # (batch, memory_max_length, depth * 2)
+        # (batch, query_max_length, depth * 2)
         queries = mx.sym.FullyConnected(data=queries,
                                         weight=self.w_q2h,
                                         bias=self.b_q2h,
                                         num_hidden=self.depth,
                                         flatten=False,
                                         name="%sq_transform" % self.prefix)
-
         return self._attend(queries,
                             keys,
                             values,
-                            queries_max_length=queries_max_length,
-                            memory_max_length=memory_max_length,
-                            lengths=memory_lengths,
-                            bias=None)
+                            bias=bias)
 
 
 class PositionalEncodings(mx.operator.CustomOp):